Allow multiple captures in the same Phlex::Html
