+++ test/rails.rb
@@ -11,4 +11,13 @@ describe Phlex::HTML do
 				%(<p>Before</p>\n\n<h1>Hello World!</h1>\n\n<article class="drop-shadow p-5 rounded">\n  <p>Start Card A</p>\n  <h3 class="font-bold">Hello from A</h3>\n  <article class="drop-shadow p-5 rounded">\n    <p>Start Card B</p>\n    <h3 class="font-bold">Hello from B</h3>\n    <p>End Card B</p>\n</article>  <p>End Card A</p>\n</article>)
 		end
 	end
+
+	with "rendered in an ERB view with mutiple captures" do
+		let(:output) { ArticlesController.render "blog" }
+
+		it "renders the correct output" do
+			expect(output).to be ==
+				%(<h3 class="font-bold"><a href="/">My blog</a></h3><article class="drop-shadow p-5 rounded"><a href="/posts/1">First Post</a></article><article class="drop-shadow p-5 rounded"><a href="/posts/2">Last Post</a></article>)
+		end
+	end
 end
